14/06/19 v5

+++ database/migrations/2019_06_11_121853_tbl_hujan.php
@@ -17,7 +17,7 @@ class TblHujan extends Migration
             $table->bigIncrements('id')->unsigned();
             $table->float('nilai_hujan_analog', 6, 2)->nullable()->default(0);
             $table->boolean('nilai_hujan_digital')->nullable()->default(0);
-            $table->tinyInteger('nilai_lembab')->nullable()->default(0);
+            $table->float('nilai_lembab', 5, 2)->nullable()->default(0);
             $table->float('nilai_suhu_c', 5, 2)->nullable()->default(0);
             $table->float('nilai_suhu_f', 5, 2)->nullable()->default(0);
             // $table->timestampTz('expired_at')->nullable();
